+++ cmake-scripts/boomerang-configure.cmake
@@ -14,7 +14,6 @@ include(CheckLibraryExists)
 include(TestBigEndian)
 
 # Boomerang configuration options
-option(BUILD_SHARED_LIBS                "Build shared or static libraries." ON)
 option(BOOMERANG_BUILD_GUI              "Build the GUI. Requires Qt5Widgets." ON)
 option(BOOMERANG_BUILD_CLI              "Build the command line interface." ON)
 option(BOOMERANG_BUILD_UNIT_TESTS       "Build the unit tests. Requires Qt5Test." OFF)
@@ -93,9 +92,3 @@ add_definitions(-DDEBUG_PARAMS=1)     #
 add_definitions(-DRECURSION_WIP=0)
 add_definitions(-DPRINT_UNION=0)      # Set to 1 to debug unions to stderr
 add_definitions(-DV9_ONLY=0)
-
-
-if (NOT BUILD_SHARED_LIBS)
-    add_definitions(-DBOOMERANG_BUILD_STATIC=1)
-endif ()
-