+++ scripts/appveyor-generate.ps1
@@ -11,7 +11,7 @@ $env:QTDIR = "$QT_BASE_DIR"
 # Install flex + bison via winflexbison
 if (!(Test-Path winflexbison.zip)) {
     [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-    if (Invoke-WebRequest -Uri http://github.com/lexxmark/winflexbison/releases/download/v2.5.16/winflexbison-2.5.16.zip -OutFile winflexbison.zip) {
+    if (Invoke-WebRequest -Uri https://github.com/lexxmark/winflexbison/releases/download/v2.5.16/winflexbison-2.5.16.zip -OutFile winflexbison.zip) {
         Write-Output "Could not download winflexbison"
         exit 1
     }
@@ -21,7 +21,7 @@ $expectedhash = "223CEBAAD0E1AF63A04F5C71F578AFB9C2233C09B5D69EA78C63F810EBD1364
 $actualhash = (Get-FileHash -Algorithm "SHA256" winflexbison.zip).hash
 
 if ($actualhash -ne $expectedhash) {
-    Write-Output "File hash does not match: Expected: $expectedhash, Actual: $actualhash" 
+    Write-Output "File hash does not match: Expected: $expectedhash, Actual: $actualhash"
     exit 1
 }
 
@@ -34,18 +34,9 @@ vcpkg install capstone:x64-windows
 
 
 # Build Visual Studio solution
-if ($env:CONFIGURATION -eq "Debug") {
-    $SHARED_LIBS = "OFF"
-}
-else {
-    $SHARED_LIBS = "ON"
-}
-
-# Build Visual Studio solution
 cmake -G "$CMAKE_GENERATOR_NAME" `
     -DCMAKE_TOOLCHAIN_FILE="c:/tools/vcpkg/scripts/buildsystems/vcpkg.cmake" `
     -DCMAKE_PREFIX_PATH="$CMAKE_PREFIX_PATH" `
     -DBOOMERANG_BUILD_UNIT_TESTS=ON `
-    -DBUILD_SHARED_LIBS="$SHARED_LIBS" `
     -DBISON_EXECUTABLE="C:/projects/boomerang/build/winflexbison/win_bison.exe" `
     -DFLEX_EXECUTABLE="C:/projects/boomerang/build/winflexbison/win_flex.exe" ..