+++ src/boomerang/core/BoomerangAPI.h
@@ -14,12 +14,22 @@
 #    ifdef _WIN32
 #        if defined(BOOMERANG_BUILD_SHARED) /* build dll */
 #            define BOOMERANG_API __declspec(dllexport)
-#        elif !defined(BOOMERANG_BUILD_STATIC) /* use dll */
+#        else /* use dll */
 #            define BOOMERANG_API __declspec(dllimport)
-#        else /* static library */
-#            define BOOMERANG_API
 #        endif
 #    else
 #        define BOOMERANG_API __attribute__((visibility("default")))
+#    endif
+#endif
+
+#ifndef BOOMERANG_PLUGIN_API
+#    ifdef _WIN32
+#        if defined(BOOMERANG_BUILD_PLUGIN) /* build plugin dll */
+#            define BOOMERANG_PLUGIN_API __declspec(dllexport)
+#        else /* use dll */
+#            define BOOMERANG_PLUGIN_API __declspec(dllimport)
+#        endif
+#    else
+#        define BOOMERANG_PLUGIN_API __attribute__((visibility("default")))
 #    endif
 #endif