+++ src/boomerang/ifc/IDecoder.h
@@ -17,6 +17,7 @@
 class Exp;
 class RTL;
 class Prog;
+class Project;
 class RTLInstDict;
 
 
@@ -27,7 +28,11 @@ class RTLInstDict;
 class IDecoder
 {
 public:
+    IDecoder(Project *) {}
     virtual ~IDecoder() = default;
+
+public:
+    virtual bool initialize(Project *project) = 0;
 
     /**
      * Decodes the machine instruction at \p pc.