+++ tests/unit-tests/boomerang/frontend/SPARCFrontEndTest.cpp
@@ -9,12 +9,12 @@
 #pragma endregion License
 #include "SPARCFrontEndTest.h"
 
+#include "boomerang-plugins/frontend/sparc/SPARCFrontEnd.h"
 
 #include "boomerang/db/BasicBlock.h"
+#include "boomerang/db/Prog.h"
 #include "boomerang/db/proc/ProcCFG.h"
 #include "boomerang/db/proc/UserProc.h"
-#include "boomerang/db/Prog.h"
-#include "boomerang/frontend/sparc/SPARCFrontEnd.h"
 #include "boomerang/ssl/RTL.h"
 #include "boomerang/util/Types.h"
 #include "boomerang/util/log/Log.h"