+++ tests/unit-tests/boomerang/ssl/type/TypeTest.cpp
@@ -14,7 +14,6 @@
 #include "boomerang/db/module/Module.h"
 #include "boomerang/db/proc/UserProc.h"
 #include "boomerang/db/signature/Signature.h"
-#include "boomerang/frontend/pentium/PentiumFrontEnd.h"
 #include "boomerang/ssl/type/ArrayType.h"
 #include "boomerang/ssl/type/CompoundType.h"
 #include "boomerang/ssl/type/IntegerType.h"
@@ -115,10 +114,13 @@ void TypeTest::testCompound()
     prog->readDefaultLibraryCatalogues();
 
     std::shared_ptr<Signature> paintSig = prog->getLibSignature("BeginPaint");
+    const SharedType paramType = paintSig->getParamType(1);
 
-    SharedType paramType = paintSig->getParamType(1);
-    QCOMPARE(paintSig->getParamType(1)->getCtype(), QString("LPPAINTSTRUCT"));
-    SharedType paintStructType = paramType->as<PointerType>()->getPointsTo();
+    QVERIFY(paramType != nullptr);
+    QCOMPARE(paramType->getCtype(), QString("LPPAINTSTRUCT"));
+
+    const SharedType paintStructType = paramType->as<PointerType>()->getPointsTo();
+    QVERIFY(paintStructType != nullptr);
     QCOMPARE(paintStructType->getCtype(), QString("PAINTSTRUCT"));
 
     // Offset 8 should have a RECT