+++ Changelog.md
@@ -9,6 +9,11 @@ v0.5.0 (in development)
 - Fixed: Unsigned integral types mistaken for signed integral types when parsing signature or symbol information.
 - Fixed: Memory leaks in parsers.
 - Feature: The x86 decoder now recognizes a larger subset of the x86 instruction set.
+- Feature: Added support for Code Generator plugins.
+- Feature: Added support for Type Recovery plugins.
+- Feature: Added support for Symbol Provider plugins.
+- Feature: Added support for Decoder plugins.
+- Feature: Added support for FrontEnd plugins.
 - Improved: Performance of decoding x86 instructions.
 - Improved: General processing of overlapped registers (not just hard-coded ones).
 - Improved: Better high level code output quality for x86 binaries due to more instructions being recognized.
@@ -22,6 +27,7 @@ v0.5.0 (in development)
 - Removed: Broken MIPS support.
 - Removed: Broken M68k support.
 - Removed: Broken HP PA/RISC support.
+- Removed: Ability to build libboomerang as a static library.
 - Technical: Improved compilation times and memory usage while compiling.
 
 v0.4.0-alpha (2018-11-11)